+++ packages/comment2/src/client/components/Artalk.ts
@@ -60,7 +60,7 @@ export default defineComponent({
         }),
       ]);
 
-      // FIXME: Typescript type issues
+      // FIXME: Types issue
       const Artalk = _Artalk as unknown as typeof _Artalk.default;
 
       loaded.value = true;
@@ -90,7 +90,7 @@ export default defineComponent({
 
     onMounted(() => {
       watch(
-        () => page.value.path,
+        () => props.identifier,
         async () => {
           try {
             artalk?.destroy();
@@ -99,14 +99,14 @@ export default defineComponent({
           }
           await initArtalk();
         },
-        { immediate: true }
+        { immediate: true },
       );
 
       watch(
         () => props.darkmode,
         (value) => {
           artalk?.setDarkMode(value);
-        }
+        },
       );
     });