+++ packages/comment2/src/node/plugin.ts
@@ -25,7 +25,7 @@ export const commentPlugin =
     // TODO: Remove this in v2 stable
     if (legacy)
       convertOptions(options as CommentPluginOptions & Record<string, unknown>);
-    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.64");
+    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.67");
 
     if (app.env.isDebug) logger.info("Options:", options);
 
@@ -68,7 +68,7 @@ export const commentPlugin =
             addViteOptimizeDepsExclude(
               bundlerOptions,
               app,
-              "artalk/dist/Artalk.es.js"
+              "artalk/dist/Artalk.es.js",
             );
             addViteSsrExternal(bundlerOptions, app, "artalk");
             break;