+++ packages/lightgallery/src/node/plugin.ts
@@ -9,7 +9,7 @@ import { CLIENT_FOLDER, PLUGIN_NAME, logger } from "./utils.js";
 export const lightgalleryPlugin =
   (options: LightGalleryOptions = {}): PluginFunction =>
   (app) => {
-    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.64");
+    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.67");
 
     if (app.env.isDebug) logger.info("Options:", options);
 
@@ -38,7 +38,7 @@ export const lightgalleryPlugin =
         addViteOptimizeDepsExclude(bundlerOptions, app, [
           "lightgallery/lightgallery.es5.js",
           ...plugins.map(
-            (name) => `lightgallery/plugins/${name}/lg-${name}.es5.js`
+            (name) => `lightgallery/plugins/${name}/lg-${name}.es5.js`,
           ),
         ]);
         addViteOptimizeDepsExclude(bundlerOptions, app, ["lightgallery"]);