+++ packages/md-enhance/src/client/components/CodeTabs.ts
@@ -8,7 +8,7 @@ import "../styles/code-tabs.scss";
 
 const codeTabStore = useStorage<Record<string, string>>(
   "VUEPRESS_CODE_TAB_STORE",
-  {}
+  {},
 );
 
 export default defineComponent({
@@ -69,6 +69,7 @@ export default defineComponent({
 
   setup(props, { slots }) {
     // index of current active item
+    // eslint-disable-next-line vue/no-setup-props-destructure
     const activeIndex = ref(props.active);
 
     // refs of the tab buttons
@@ -112,7 +113,7 @@ export default defineComponent({
     const getInitialIndex = (): number => {
       if (props.tabId) {
         const valueIndex = props.data.findIndex(
-          ({ id }) => codeTabStore.value[props.tabId] === id
+          ({ id }) => codeTabStore.value[props.tabId] === id,
         );
 
         if (valueIndex !== -1) return valueIndex;
@@ -132,7 +133,7 @@ export default defineComponent({
 
             if (index !== -1) activeIndex.value = index;
           }
-        }
+        },
       );
     });
 
@@ -164,9 +165,9 @@ export default defineComponent({
                     onKeydown: (event: KeyboardEvent) =>
                       keyboardHandler(event, index),
                   },
-                  slots[`title${index}`]({ value: id, isActive })
+                  slots[`title${index}`]({ value: id, isActive }),
                 );
-              })
+              }),
             ),
             props.data.map(({ id }, index) => {
               const isActive = index === activeIndex.value;
@@ -179,7 +180,7 @@ export default defineComponent({
                   role: "tabpanel",
                   "aria-expanded": isActive,
                 },
-                slots[`tab${index}`]({ value: id, isActive })
+                slots[`tab${index}`]({ value: id, isActive }),
               );
             }),
           ])