+++ packages/md-enhance/src/node/plugin.ts
@@ -50,6 +50,7 @@ import {
   echarts,
   flowchart,
   getTSPlaygroundPreset,
+  getUnoPlaygroundPreset,
   getVuePlaygroundPreset,
   hint,
   mermaid,
@@ -74,22 +75,22 @@ import { PLUGIN_NAME, logger } from "./utils.js";
 export const mdEnhancePlugin =
   (
     options: MarkdownEnhanceOptions = { gfm: true },
-    legacy = true
+    legacy = true,
   ): PluginFunction =>
   (app) => {
     // TODO: Remove this in v2 stable
     if (legacy)
       convertOptions(
-        options as MarkdownEnhanceOptions & Record<string, unknown>
+        options as MarkdownEnhanceOptions & Record<string, unknown>,
       );
 
-    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.64");
+    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.67");
 
     if (app.env.isDebug) logger.info("Options:", options);
 
     const getStatus = (
       key: keyof MarkdownEnhanceOptions,
-      gfm = false
+      gfm = false,
     ): boolean =>
       key in options ? Boolean(options[key]) : (gfm && options.gfm) || false;
 
@@ -115,7 +116,7 @@ export const mdEnhancePlugin =
 
     const { enabled: linksCheckEnabled, isIgnoreLink } = getLinksCheckStatus(
       app,
-      options
+      options,
     );
 
     const katexOptions: KatexOptions<MarkdownEnv> = {
@@ -137,7 +138,7 @@ export const mdEnhancePlugin =
           logger.warn(
             `Found unicode character ${token.text} inside tex${
               filePathRelative ? ` in ${colors.cyan(filePathRelative)}` : ""
-            }. You should use ${colors.magenta(`\\text{${token.text}}`)}`
+            }. You should use ${colors.magenta(`\\text{${token.text}}`)}`,
           );
         else
           logger.warn(
@@ -145,7 +146,7 @@ export const mdEnhancePlugin =
               filePathRelative
                 ? `\nFound in ${colors.cyan(filePathRelative)}`
                 : ""
-            }`
+            }`,
           );
       },
       ...(isPlainObject(options.katex) ? options.katex : {}),
@@ -205,7 +206,7 @@ export const mdEnhancePlugin =
           addViteOptimizeDepsExclude(
             bundlerOptions,
             app,
-            "chart.js/auto/auto.mjs"
+            "chart.js/auto/auto.mjs",
           );
           addViteSsrExternal(bundlerOptions, app, "chart.js");
         }
@@ -230,7 +231,7 @@ export const mdEnhancePlugin =
             "reveal.js/dist/reveal.esm.js",
             "reveal.js/plugin/markdown/markdown.esm.js",
             ...revealPlugins.map(
-              (plugin) => `reveal.js/plugin/${plugin}/${plugin}.esm.js`
+              (plugin) => `reveal.js/plugin/${plugin}/${plugin}.esm.js`,
             ),
           ]);
 
@@ -261,7 +262,7 @@ export const mdEnhancePlugin =
         if (imgMarkEnable)
           md.use(
             imgMark,
-            isPlainObject(options.imgMark) ? options.imgMark : {}
+            isPlainObject(options.imgMark) ? options.imgMark : {},
           );
 
         if (getStatus("imgSize")) md.use(imgSize);
@@ -351,6 +352,8 @@ export const mdEnhancePlugin =
               md.use(playground, getTSPlaygroundPreset(config.ts || {}));
             else if (preset === "vue")
               md.use(playground, getVuePlaygroundPreset(config.vue || {}));
+            else if (preset === "unocss")
+              md.use(playground, getUnoPlaygroundPreset(config.unocss || {}));
             else if (isPlainObject(preset)) md.use(playground, preset);
           });
         }