+++ packages/reading-time2/src/node/plugin.ts
@@ -11,7 +11,7 @@ import type { ReadingTime } from "../shared/index.js";
 export const readingTimePlugin =
   (options: ReadingTimeOptions): PluginFunction =>
   (app) => {
-    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.64");
+    checkVersion(app, PLUGIN_NAME, "2.0.0-beta.67");
 
     if (app.env.isDebug) logger.info("Options:", options);
 
@@ -30,7 +30,7 @@ export const readingTimePlugin =
       extendsPage: (page: Page<{ readingTime?: ReadingTime }>): void => {
         page.data.readingTime = getReadingTime(
           page.content,
-          options.wordPerMinute || 300
+          options.wordPerMinute || 300,
         );
       },
     };