+++ packages/theme/src/client/modules/info/components/PageViewInfo.ts
@@ -1,4 +1,3 @@
-import { withBase } from "@vuepress/client";
 import { isString } from "@vuepress/shared";
 import { useMutationObserver } from "@vueuse/core";
 import type { VNode } from "vue";
@@ -46,7 +45,7 @@ export default defineComponent({
 
         if (count && !isNaN(Number(count))) pageViews.value = Number(count);
       },
-      { childList: true }
+      { childList: true },
     );
 
     return (): VNode | null =>
@@ -71,11 +70,11 @@ export default defineComponent({
                   /** visitorID */
                   "data-path": isString(props.pageview)
                     ? props.pageview
-                    : withBase(route.path),
+                    : route.path,
                 },
-                "..."
+                "...",
               ),
-            ]
+            ],
           )
         : null;
   },