+++ packages/theme/src/client/utils/resolveLinkInfo.ts
@@ -14,14 +14,15 @@ import { ArticleInfoType } from "../../shared/index.js";
 export const resolveLinkInfo = (
   router: Router,
   item: string,
-  preferFull = false
+  preferFull = false,
 ): AutoLinkOptions => {
-  const encodedPath = encodeURI(item);
-  let result = resolveRouteWithRedirect(router, inferRouteLink(encodedPath));
+  let result = resolveRouteWithRedirect(
+    router,
+    inferRouteLink(encodeURI(item)),
+  );
 
   // the inferred path may be wrong, so we need to resolve the original path
-  if (result.name === "404")
-    result = resolveRouteWithRedirect(router, encodedPath);
+  if (result.name === "404") result = resolveRouteWithRedirect(router, item);
 
   const { fullPath, meta, name } = result;