+++ packages/theme/src/node/plugins/blog/blog.ts
@@ -23,7 +23,7 @@ export const getBlogPlugin = (
   app: App,
   themeData: ThemeData,
   options?: BlogPluginOptions | boolean,
-  hotReload = false
+  hotReload = false,
 ): Plugin | null => {
   if (!options) return null;
 
@@ -91,5 +91,6 @@ export const getBlogPlugin = (
 
     hotReload,
     ...("hotReload" in blogOptions ? { hotReload: blogOptions.hotReload } : {}),
+    ...("slugify" in blogOptions ? { slugify: blogOptions.slugify } : {}),
   });
 };